+++ src/vivarium/examples/disease_model/__init__.py
@@ -1,4 +1,6 @@
-# mypy: ignore-errors
+from pathlib import Path
+
+from vivarium import InteractiveContext
 from vivarium.examples.disease_model.disease import (
     DiseaseModel,
     DiseaseState,
@@ -12,15 +14,11 @@ from vivarium.examples.disease_model.population import BasePopulation
 from vivarium.examples.disease_model.risk import Risk, RiskEffect
 
 
-def get_model_specification_path():
-    from pathlib import Path
-
+def get_model_specification_path() -> str:
     p = Path(__file__).parent / "disease_model.yaml"
     return str(p)
 
 
-def get_disease_model_simulation():
-    from vivarium import InteractiveContext
-
+def get_disease_model_simulation() -> InteractiveContext:
     p = get_model_specification_path()
     return InteractiveContext(p)