+++ CHANGELOG.rst
@@ -1,3 +1,9 @@
+**3.3.15 - 04/14/25**
+
+  - Type-hinting: Fix mypy errors in src/vivarium/interface/cli.py
+  - Type-hinting: Fix mypy errors in tests/interface/test_cli.py
+  - Type-hinting: Fix mypy erros in tests/framework/test_configuration.py
+
 **3.3.14 - 03/31/25**
 
   - Feature: Add classmethod to InteractiveContext to create instance from a backup